% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdpackage-package.R
\docType{package}
\name{jdpackage-package}
\alias{jdpackage}
\alias{jdpackage-package}
\title{jdpackage: What the Package Does (One Line, Title Case)}
\description{
Test how packaging works in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Juke34/jdpackage}
  \item Report bugs at \url{https://github.com/Juke34/jdpackage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
